+++ CHANGELOG.md
@@ -8,7 +8,7 @@ this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 
 ### Added
 
-- Support RADIUS protocol for sending giganto logs.
+- Support `RADIUS`, `malformed_dns` protocol for sending giganto logs.
 
 ### Changed